video refactor
